Refactored (java-application): moved JsonValideException.java out of the impl-package, cause its more a interface than a implementation
